refactored start consultation endpoint
