+++ VideoAPI/VideoApi.DAL/VideoApi.DAL.csproj
@@ -30,6 +30,5 @@
   <ItemGroup>
     <ProjectReference Include="..\VideoApi.Common\VideoApi.Common.csproj" />
     <ProjectReference Include="..\VideoApi.Domain\VideoApi.Domain.csproj" />
-    <ProjectReference Include="..\VideoApi.Services\VideoApi.Services.csproj" />
   </ItemGroup>
 </Project>
