+++ VideoAPI/VideoApi.Services/Contracts/IConsultationService.cs
@@ -1,14 +1,14 @@
-using System;
 using System.Threading.Tasks;
-using VideoApi.Domain.Enums;
-using VideoApi.Services.Kinly;
+using Microsoft.AspNetCore.Mvc;
+using VideoApi.Contract.Requests;
+using VideoApi.Domain;
 
 namespace VideoApi.Services.Contracts
 {
     public interface IConsultationService
     {
-        Task<CreateConsultationRoomResponse> CreateConsultationRoomAsync(string virtualCourtRoomId, CreateConsultationRoomParams createConsultationRoomParams);
+        Task<Room> GetAvailableConsultationRoomAsync(StartConsultationRequest request);
 
-        Task TransferParticipantAsync(Guid conferenceId, Guid participantId, string fromRoom, string toRoom);
+        Task<IActionResult> TransferParticipantToConsultationRoomAsync(StartConsultationRequest request, Room room);
     }
 }
