+++ VideoAPI/VideoApi.Services/VideoApi.Services.csproj
@@ -6,6 +6,7 @@
 
     <ItemGroup>
       <PackageReference Include="Azure.Storage.Blobs" Version="12.4.2" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="2.2.0" />
       <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.9" />
       <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.1.1" />
       <PackageReference Include="Polly" Version="7.2.0" />
